extract SlackService from ISlackService file
